+++ CHANGELOG.md
@@ -14,6 +14,7 @@ develop
 ### Fixes
 
 - Fix an issue where UNIX scripts could fail getting the content of the document.
+- Address an issue where the application could crash on document saving or text replacement.