+++ CHANGELOG.md
@@ -16,7 +16,7 @@ develop
 - Fix an issue where UNIX scripts could fail getting the content of the document.
 - Fix an issue where font change in the preferences pane could be ignored.
 - Workaround an issue where word suggestion in the Touch Bar cannot insert a word starts with a symbol correctly.
-- Address an issue where the application could crash on document saving or text replacement.
+- Workaround an issue where the application could crash on document auto-saving.